requirements/shared.txt
# Shared dependencies (included by all services)
asyncpg==0.29.0
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
